add backend sql in admin_panel

+++ sql_query.py
@@ -11,16 +11,35 @@ class MySQLQuery:
         )
         self.cursor = self.db.cursor()
 
-    # 1. Add Student
-    def add_student(self, student_data):
+
+    def add_users(self, user_info):
         try:
-            sql = """
-            INSERT INTO students 
-            (username, s_name, class, roll, section, grade, pn_number, address, tution_fee, paid_fee) 
-            VALUES (%s, %s, %s, %s, %s, %s, %s, %s, 1000, 0);
-            """
-            self.cursor.execute(sql, student_data)
+            sql = """INSERT INTO users (username, password, role)
+            VALUES (%s, %s, %s);"""
+
+            values = (user_info['username'], user_info['pass'], user_info['role'] )
+            
+            self.cursor.execute(sql, values)
             self.db.commit()
+
+            return True
+        except mysql.connector.IntegrityError as e:
+            return f"Error : {e}"
+        except Exception as e:
+            return f"Error : {e}"
+        pass
+
+    # 1. Add Student
+    def add_student(self, student_data: dict):
+        try:
+            sql = """INSERT INTO students (username, s_name, class, roll, section, grade, pn_number, address, tution_fee, paid_fee)
+            VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s);"""
+
+            values = (student_data['username'], student_data['s_name'], student_data['class'], student_data['roll'], student_data['section'], student_data['grade'], student_data['pn_number'], student_data['address'], student_data['tution_fee'], student_data['paid_fee'] )
+            
+            self.cursor.execute(sql, values)
+            self.db.commit()
+
             return True
         except mysql.connector.IntegrityError as e:
             return f"Error : {e}"
@@ -30,49 +49,48 @@ class MySQLQuery:
 
     # 2. Find Student
     def find_student(self, username):
-        self.cursor.execute("SELECT * FROM students WHERE username = %s", (username,))
-        return self.cursor.fetchone()
+        self.cursor.execute("SELECT * FROM students WHERE username = %s;", (username,))
+        f_s = self.cursor.fetchone()
+        return f_s
+        
+
 
     # 3. Delete Student
     def delete_student(self, username):
         try:
-            self.cursor.execute("DELETE FROM students WHERE username = %s", (username,))
-            return True
+            ck = self.find_student(username)
+            if ck is not None:
+                self.cursor.execute("DELETE FROM students WHERE username = %s;", (username,))
+                self.db.commit()
+                return True
+            else:
+                return "❌ Student Not Found"
+            
         except Exception as e:
-            print(f"[Error] Failed to delete student: {e}")
-            return False
+            return f"❌ [Error] Failed to delete student: {e}"
 
-    # 4. Student Info
-    def student_info(self, username):
-        self.cursor.execute("SELECT name, grade, section, roll FROM students WHERE username = %s", (username,))
-        return self.cursor.fetchone()
-
-    # 5. All Students
+    # 4. All Students
     def all_students(self):
-        self.cursor.execute("SELECT * FROM students")
+        self.cursor.execute("SELECT * FROM students;")
         return self.cursor.fetchall()
 
-    # 6. Student Routine
+    # 5. Student Routine
     def student_routine(self, username):
         self.cursor.execute("SELECT subject_name, class, section, class_start_time, class_end_time FROM subject WHERE username = %s", (username,))
         return self.cursor.fetchall()
 
-    # 7. Add Teacher
+    # 6. Add Teacher
     def add_teacher(self, teacher_data: dict):
-        required_fields = ['username', 'name', 'phone', 'address']
-        missing_fields = [field for field in required_fields if field not in teacher_data]
-        if missing_fields:
-            print(f"[Error] Missing teacher fields: {', '.join(missing_fields)}")
-            return False
-
+        
         try:
-            sql = """INSERT INTO teachers (username, name, phone, address)
+            sql = """INSERT INTO teacher (username, t_name, t_pn_number, t_address)
                      VALUES (%s, %s, %s, %s)"""
             values = (
                 teacher_data['username'], teacher_data['name'],
                 teacher_data['phone'], teacher_data['address']
             )
             self.cursor.execute(sql, values)
+            self.db.commit()
             return True
         except mysql.connector.IntegrityError as e:
             print(f"[Error] Failed to add teacher: {e}")
@@ -81,36 +99,73 @@ class MySQLQuery:
             print(f"[Error] Unexpected error: {e}")
             return False
 
-    # 8. Find Teacher
+    # 7. Find Teacher
     def find_teacher(self, username):
-        self.cursor.execute("SELECT * FROM teachers WHERE username = %s", (username,))
-        return self.cursor.fetchone()
+        self.cursor.execute("SELECT * FROM teacher WHERE username = %s", (username,))
+        f_t = self.cursor.fetchone()
+        return f_t
 
-    # 9. Delete Teacher
+    # 8. Delete Teacher
     def delete_teacher(self, username):
         try:
-            self.cursor.execute("DELETE FROM teachers WHERE username = %s", (username,))
-            return True
+            ck = self.find_teacher(username)
+            if ck is not None:
+                self.cursor.execute("DELETE FROM teacher WHERE username = %s;", (username,))
+                self.db.commit()
+                return True
+            else:
+                return "❌ Teacher Not Found"
+            
         except Exception as e:
-            print(f"[Error] Failed to delete teacher: {e}")
-            return False
+            return f"❌ [Error] Failed to delete student: {e}"
 
-    # 10. Teacher Info
+    # 9. Teacher Info
     def teacher_info(self, username):
-        self.cursor.execute("SELECT name, phone FROM teachers WHERE username = %s", (username,))
+        self.cursor.execute("SELECT name, phone FROM teacher WHERE username = %s", (username,))
         return self.cursor.fetchone()
 
-    # 11. All Teachers
+    # 10. All Teachers
     def all_teachers(self):
-        self.cursor.execute("SELECT * FROM teachers")
+        self.cursor.execute("SELECT * FROM teacher;")
         return self.cursor.fetchall()
 
-    # 12. Teacher Routine
+    # 11. Teacher Routine
     def teacher_routine(self, username):
         self.cursor.execute("SELECT subject_name, class, section, class_start_time, class_end_time FROM subject WHERE username = %s", (username,))
         return self.cursor.fetchall()
+    
+    def add_subject(self, sub_info: dict):
 
-    # 13. Student Fee
+        try:
+            sql = """INSERT INTO subjects (sub_name, username, t_name, class, class_start_time, class_end_time, section)
+                     VALUES (%s, %s, %s, %s, %s, %s, %s);"""
+            values = (
+                sub_info['subject'],sub_info['username'], sub_info['name'], 
+                sub_info['class'], sub_info['start_t'], sub_info['end_t'],sub_info['section']
+            )
+            self.cursor.execute(sql, values)
+            self.db.commit()
+            return True
+        except mysql.connector.IntegrityError as e:
+            print(f"[Error] Failed to add teacher: {e}")
+            return False
+        except Exception as e:
+            print(f"[Error] Unexpected error: {e}")
+            return False
+
+
+    # 12. Student Fee
     def student_fee(self, username):
         self.cursor.execute("SELECT tuition_fee, paid_fee, (tuition_fee - paid_fee) AS due_fee FROM students WHERE username = %s", (username,))
         return self.cursor.fetchone()
+    
+    def update_fees(self, username, tution_fee, paid_fee):
+        try:
+            self.cursor.execute("UPDATE students SET tution_fees = %s, paid_fees = %s WHERE = %s", (tution_fee, paid_fee, username,))
+            self.db.commit()
+            return True
+        except:
+            return False
+
+
+# UPDATE table_name SET column_name = new_value WHERE condition;
